refactor(schemas): standardize file naming convention
