+++ server/src/infra/http/schemas/auth.ts
