+++ server/src/infra/http/schemas/pharmacy.ts
